updated so opentree sends out ete3 not dendropy tree
