+++ topiary/__init__.py
@@ -3,7 +3,7 @@
 from .topiary import ncbi_blast_xml_to_df
 from .topiary import remove_redundancy, reverse_blast
 from .topiary import load_fasta, write_fasta, write_phy
-from .opentree import get_ott_id, get_species_tree, build_species_corrected_gene_tree
+from .opentree import get_ott_id, get_species_tree
 
 from . import ncbi
 from . import util
